Update Block IDs and Tracking Events to fix tracking problems
Changes to the onboarding flow mean that the tracking events weren't all firing, and some were being duplicated. This updates the Blockly block IDs and simplifies the trackingEvents structure to make the events easier to track, and more robust – no longer depending on the block values, but only their IDs.