Add password reset and username request facilities on `kano-auth` component
Email addresses are not unique to usernames, so users need to be able to request all usernames associated with their email addresses, and request password resets for user accounts individually. Simple forms on the `kano-auth` component make post requests to a new and updated route on the api to trigger password-rest and username-request emails.

Trello cards:
https://trello.com/c/iMFJqmQN/1194-2-forgot-username-button
https://trello.com/c/tEe13RUb/1193-5-password-reset-based-on-email-address-not-being-unique