Add tracking to make use of new API tracking schemas, and allow tracking within the Electron app
* Renames older tracking behavior to nest this under the new `Tracking` object, and give this a Google-specific name
* Update events being fired to make them GA-specific and be caught by the newly-renamed behavior
* Add new `KanoCodeBehavior` for tracking to provide support for Kano Code pageviews and challenge events
* Update the `kano-view-behavior` and `kano-sharing-behavior` to update previous tracking events and add in new ones
* Add new tracking behavior to `kano-app`

The largest changes are on `kano-view-story` to add new `_trackRewards` function to begin sorting rewards and firing the correct events for these. As well as `kano-sharing-behavior` to add a new `remixedShare` property to keep track of whether shares have been remixed from other shares, and supply the required information about these for tracking events. The sharing flow has also been slightly updated with `_shareApp` take an `options` object so that tracking events can be suppressed when the link is being requested from the API, and only fired when users have chosen to save to Kano World.

Depends on:
https://github.com/KanoComputing/kano-web-api/pull/394
https://github.com/KanoComputing/web-components/pull/107

and related to:
https://github.com/KanoComputing/kano2-app/pull/221