Use getItem for localStorage progress, rather than accessing the progress key directly
Prevents issues in the app, where Kano Code is accessing the `progress` key directly, and bypassing the `localStorage` overrides. This means that existing progress from Cross Storage is being accessed instead of the local copy, and each user is being updated with the previous user's progress.

Trello card: https://trello.com/c/oTZWItox/527-only-finished-one-challenge-but-it-says-i-have-completed-more