"Back to challenges" triggers `window.history.back()`
The `view-device` page now requires a `serialNumber` query argument if you want to select a device from that type (specified by the `product` query argument).

Since the `Kano.MakeApps.config.PROJECTS_URL` won't have this information and we have a router in place, the safest solution would be to use what the "back" button is using which is `window.history.back()` in order to go back to the challenges page. We can (hopefully) assume that if you are on the end of a challenge, you probably were on the challenges page.