Do not run crisper on the main index file
and check for `window.Kano` and `window.Kano.MakeApps` before declaring.

Currently, the build process is crisping the main index file wen copying all the files to the `.tmp` directory, so when the `htmlReplace` runs, we're ending up with duplicate declarations of the `config`. Excluding the crisping of the index file allows the `htmlReplace` ot actually replace the stand-in scripts, and ensure that values like the `VERSION` are not being overwritten.