Fixing a race in the build
Some weird race condition was causing inline scripts to be removed from
the index.html file which broke the build. Moving the copy-index after
the parallel tasks seems to fix the issue for now, but I couldn't
actually track it down to exactly what was causing the removal (none of
the parallel tasks seem to affect the build).